to float

+++ models/bert/bert_classifier.py
@@ -173,10 +173,10 @@ class BertClassifierWithMetadataXS(BertPreTrainedModel):
         # mask = (r_input_ids == 0) # (batch size, seq length)
 
         # Don't account for completely zeroed out sub-batches
-        mask = (attn_mask.view(batch_size, batch_increase_factor, metadata_len).sum(dim=2) > 0).unsqueeze(2)
+        mask = (attn_mask.view(batch_size, batch_increase_factor, metadata_len).sum(dim=2) > 0).type(torch.float).unsqueeze(2)
         pooled_output = pooled_output.view(batch_size, batch_increase_factor, -1)
         pooled_output = pooled_output * mask
-        pooled_output = torch.sum(pooled_output, dim=1) / torch.sum(mask.squeeze(2).type(torch.long), dim=1)
+        pooled_output = torch.sum(pooled_output, dim=1) / torch.sum(mask.squeeze(2).type(torch.float), dim=1)
 
         # mask = (mask.view(batch_size, batch_increase_factor, metadata_len).sum(dim=2) == metadata_len).unsqueeze(2) # bs, batch_increase_factor
         # pooled_output = pooled_output.view(batch_size, batch_increase_factor, -1)
